# platform/swaggo/swagger.yaml
basePath: /flighthours/api/v1
definitions:
...
  handlers.CreateTailNumberRequest:
    properties:
      aircraft_model_id:
        type: string
      airline_id:
        type: string
      tail_number:
        type: string
    required:
    - aircraft_model_id
    - airline_id
    - tail_number
    type: object
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
...
  handlers.TailNumberListResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      registrations:
        items:
          $ref: '#/definitions/handlers.TailNumberResponse'
        type: array
      total:
        type: integer
    type: object
  handlers.TailNumberResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      aircraft_model_id:
        type: string
      airline_id:
        type: string
      airline_name:
        description: Aerolínea (denormalized)
        type: string
      id:
        type: string
      model_name:
        description: Modelo (denormalized)
        type: string
      tail_number:
        description: Numero de Matrícula
        type: string
    type: object
...
  handlers.UpdateTailNumberRequest:
    properties:
      aircraft_model_id:
        type: string
      airline_id:
        type: string
      tail_number:
        type: string
    required:
    - aircraft_model_id
    - airline_id
    - tail_number
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@champion19.com
    name: Champion19 Support
  description: API para gestión de horas de vuelo de pilotos y tripulantes
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flighthours API
  version: "1.0"
paths:
...
  /tail-numbers:
    get:
      description: Returns a list of all aircraft registrations with optional filters
      parameters:
      - description: Filter by airline ID
        in: query
        name: airline_id
        type: string
      - description: Filter by tail number
        in: query
        name: tail_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TailNumberListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all aircraft registrations
      tags:
      - Tail Numbers
    post:
      consumes:
      - application/json
      description: Creates a new aircraft registration (tail number must be unique)
      parameters:
      - description: Aircraft registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTailNumberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.TailNumberResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new aircraft registration
      tags:
      - Tail Numbers
  /tail-numbers/{id}:
    put:
      consumes:
      - application/json
      description: Updates an aircraft registration by ID (accepts both UUID and obfuscated
        ID)
      parameters:
      - description: Tail Number ID (obfuscated ID)
        in: path
        name: id
        required: true
        type: string
      - description: Aircraft registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTailNumberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TailNumberResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing aircraft registration
      tags:
      - Tail Numbers
  /tail-numbers/{plate}:
    get:
      consumes:
      - application/json
      description: Returns aircraft registration information by its tail number (e.g.
        HK-5432)
      parameters:
      - description: Tail number (e.g. HK-5432)
        in: path
        name: plate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TailNumberResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get aircraft registration by tail number
      tags:
      - Tail Numbers
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"